M2-4378:  moved logic into a separate hook
